+++ packages/builder/src/settings/pages/ai/completion-models/AIConfigList.svelte
