+++ packages/builder/src/settings/pages/ai/completion-models/AILogo.svelte
