+++ packages/builder/src/settings/pages/ai/completion-models/SetDefaultAIConfigButton.svelte
