+++ packages/builder/src/settings/pages/index.ts
@@ -14,7 +14,7 @@ import EmailTemplatesPage from "@/settings/pages/email/EmailTemplates.svelte"
 import EmailTemplatePage from "@/settings/pages/email/Template.svelte"
 import AuditLogsPage from "@/settings/pages/auditLogs/index.svelte"
 import AIPage from "@/settings/pages/ai/index.svelte"
-import AIConfigsPage from "@/settings/pages/ai/configs.svelte"
+import AIConfigsPage from "@/settings/pages/ai/completion-models/index.svelte"
 import AuthPage from "@/settings/pages/auth/index.svelte"
 import OrgPage from "@/settings/pages/organisation.svelte"
 import BrandingPage from "@/settings/pages/branding.svelte"
@@ -38,7 +38,7 @@ import Recaptcha from "@/settings/pages/recaptcha.svelte"
 
 // AI config
 import KnowledgeBasesPage from "@/settings/pages/ai/knowledge-bases/index.svelte"
-import AIConfigForm from "@/settings/pages/ai/AIConfigForm.svelte"
+import AIConfigForm from "@/settings/pages/ai/completion-models/AIConfigForm.svelte"
 import KnowledgeBaseForm from "./ai/knowledge-bases/KnowledgeBaseForm.svelte"
 import EmbeddingModelForm from "./ai/knowledge-bases/EmbeddingModelForm.svelte"
 import VectorDatabaseForm from "./ai/knowledge-bases/VectorDatabaseForm.svelte"
